# openapi.yaml
---
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:9000
    description: Generated server url
paths:
  "/api/token/items/search/":
    get:
      tags:
        - token-item-controller
      operationId: search
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/token/items/search/{text}":
    get:
      tags:
        - token-item-controller
      operationId: search_1
      parameters:
        - name: text
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/auth/signin":
    post:
      tags:
        - jwt-auth-controller
      operationId: signin
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthenticationRequest"
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/items/search":
    post:
      tags:
        - jwt-item-controller
      operationId: search_2
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Search"
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/items/search/{text}":
    get:
      tags:
        - jwt-item-controller
      operationId: search_3
      parameters:
        - name: text
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/items/search/":
    get:
      tags:
        - jwt-item-controller
      operationId: searchAll
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/users/search/{text}":
    get:
      tags:
        - jwt-user-controller
      operationId: search_4
      parameters:
        - name: text
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/users/search/bad/{text}":
    get:
      tags:
        - jwt-user-controller
      operationId: searchCrappy
      parameters:
        - name: text
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/jwt/users/search/":
    get:
      tags:
        - jwt-user-controller
      operationId: searchAll_1
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/basic/items/search/":
    get:
      tags:
        - basic-auth-item-controller
      operationId: search_5
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
  "/api/basic/items/search/{text}":
    get:
      tags:
        - basic-auth-item-controller
      operationId: search_6
      parameters:
        - name: text
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: default response
          content:
            "*/*":
              schema:
                type: string
components:
  schemas:
    AuthenticationRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        tenant:
          type: string
    Search:
      type: object
      properties:
        searchText:
          type: string